vsac_variables.py better refactoring.
