+++ tests/test_aspect.py
@@ -1,13 +1,13 @@
 import unittest
 from irae.variable.aspect import Aspect, AspectMap, AspectKey
-from irae.variable.aspect import Proc, Variable, Valueset
-from irae.variable.vsac_variables_defined import Dx, Rx, Lab, Proc
+from irae.variable.aspect import Procedures, Variable, Valueset
+from irae.variable.vsac_variables_defined import Diagnoses, Medications, Labs, Procedures
 from irae.variable.vsac_variables_defined import get_aspect_map
 
 class TestAspect(unittest.TestCase):
 
     def get_proc_variable_list(self):
-        return Proc([
+        return Procedures([
             Variable('Nephrectomy', {
                 'sct': '2.16.840.1.113762.1.4.1248.200',
                 'icd10pcs': '2.16.840.1.113762.1.4.1248.4'}),
@@ -15,7 +15,7 @@ class TestAspect(unittest.TestCase):
                 'services': '2.16.840.1.113883.3.464.1003.109.12.1013'})])
 
     def get_proc_variable_dict(self):
-        return Proc({
+        return Procedures({
             'Nephrectomy': {
                 'sct': '2.16.840.1.113762.1.4.1248.200',
                 'icd10pcs': '2.16.840.1.113762.1.4.1248.4'},
@@ -38,6 +38,6 @@ class TestAspect(unittest.TestCase):
         self.assertTrue(AspectKey.proc.name in aspect_json.keys())
 
         self.assertTrue(len(aspect_map.diagnosis.variable_list) > 1)
-        self.assertTrue(len(aspect_map.medication.variable_list) > 1)
-        self.assertTrue(len(aspect_map.lab.variable_list) > 1)
-        self.assertTrue(len(aspect_map.procedure.variable_list) > 1)
+        self.assertTrue(len(aspect_map.medications.variable_list) > 1)
+        self.assertTrue(len(aspect_map.labs.variable_list) > 1)
+        self.assertTrue(len(aspect_map.procedures.variable_list) > 1)
